+++ lib/arduino_ci/host.rb
@@ -1,0 +1,19 @@
+module ArduinoCI
+
+  # Tools for interacting with the host machine
+  class Host
+    # Cross-platform way of finding an executable in the $PATH.
+    # via https://stackoverflow.com/a/5471032/2063546
+    #   which('ruby') #=> /usr/bin/ruby
+    def self.which(cmd)
+      exts = ENV['PATHEXT'] ? ENV['PATHEXT'].split(';') : ['']
+      ENV['PATH'].split(File::PATH_SEPARATOR).each do |path|
+        exts.each do |ext|
+          exe = File.join(path, "#{cmd}#{ext}")
+          return exe if File.executable?(exe) && !File.directory?(exe)
+        end
+      end
+      nil
+    end
+  end
+end